Move: UserInputコンポーネントとコンテナをそれぞれmoleculesに移動
